refactor: application 부분과 web server 부분 분리
